Initial UI for `Social - Discover Your Gym Mates` page
