+++ frontend/TritonFitReactNative/App.tsx
@@ -4,105 +4,25 @@ import * as React from 'react';
 import { View, Text, TouchableOpacity, StyleSheet, TextInput } from 'react-native';
 import { NavigationContainer, useNavigation } from '@react-navigation/native';
 import { NativeStackScreenProps, createNativeStackNavigator } from '@react-navigation/native-stack';
+import HomeScreen from './screens/HomeScreen.tsx';
+import LoginScreen from './screens/LoginScreen.tsx';
+import SocialScreen from './screens/SocialScreen.tsx';
 
 // Define the type for the stack navigator
-type RootStackParamList = {
+export type RootStackParamList = {
   Home: undefined;
   Login: undefined;
+  Social: undefined;
 };
 
 const Stack = createNativeStackNavigator<RootStackParamList>();
-
-// Fix useNavigation typing
-import { NavigationProp } from '@react-navigation/native';
-
-function HomeScreen() {
-  const navigation = useNavigation<NavigationProp<RootStackParamList>>();
-
-  return (
-    <View style={styles.container}>
-      <Text style={styles.title}>TritonFit</Text>
-
-      <TouchableOpacity style={styles.button} onPress={() => navigation.navigate('Login')}>
-        <Text style={styles.buttonText}>Sign In</Text>
-      </TouchableOpacity>
-
-      <TouchableOpacity style={styles.button} >
-        <Text style={styles.buttonText}>Create Account</Text>
-      </TouchableOpacity>
-    </View>
-  );
-}
-
-function LoginScreen() {
-  return (
-    <View style={styles.container}>
-      <View style={styles.login}>
-        <Text style={styles.textInputLabel}>Username</Text>
-        <TextInput style={styles.textInput}></TextInput>
-
-        <Text style={styles.textInputLabel}>Password</Text>
-        <TextInput style={styles.textInput} secureTextEntry={true}></TextInput>
-
-        <TouchableOpacity style={styles.button}>
-          <Text style={styles.buttonText}>Sign in</Text>
-        </TouchableOpacity>
-      </View>
-    </View>    
-  )
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#182B49',
-    paddingBottom: 40,
-  },
-  title: {
-    fontSize: 48,
-    fontWeight: 'bold',
-    marginBottom: 20,
-    color: 'white',
-  },
-  button: {
-    backgroundColor: '#C69214',
-    paddingTop: 8,
-    paddingRight: 10,
-    paddingBottom: 8,
-    paddingLeft: 10,
-    borderRadius: 8,
-    marginBottom: 30, // Adds spacing between buttons
-    width: 190,
-  },
-  buttonText: {
-    color: 'white',
-    fontSize: 18,
-    fontWeight: 'bold',
-    textAlign: 'center',
-  },
-  textInputLabel: {
-    color: 'white',
-    alignSelf: 'flex-start',
-    fontSize: 16,
-  },
-  textInput: {
-    color: 'white',
-    borderWidth: 1,
-    borderColor: 'white',
-    borderRadius: 5,
-  },
-  login: {
-    width: 190,
-  }
-});
 
 function RootStack() {
   return (
     <Stack.Navigator initialRouteName="Home">
       <Stack.Screen name="Home" component={HomeScreen} />
       <Stack.Screen name="Login" component={LoginScreen} />
+      <Stack.Screen name="Social" component={SocialScreen} />
     </Stack.Navigator>
   );
 }
