+++ frontend/TritonFitReactNative/screens/LoginScreen.tsx
@@ -1,0 +1,74 @@
+// screens/LoginScreen.tsx
+import React from "react";
+import { View, Text, TouchableOpacity, Button, StyleSheet, TextInput } from "react-native";
+import { useNavigation } from '@react-navigation/native';
+import { NavigationProp } from '@react-navigation/native';
+import { RootStackParamList } from '../App.tsx';
+
+const LoginScreen = () => {
+  const navigation = useNavigation<NavigationProp<RootStackParamList>>();
+  
+  return (
+    <View style={styles.container}>
+      <View style={styles.login}>
+        <Text style={styles.textInputLabel}>Username</Text>
+        <TextInput style={styles.textInput}></TextInput>
+
+        <Text style={styles.textInputLabel}>Password</Text>
+        <TextInput style={styles.textInput} secureTextEntry={true}></TextInput>
+
+        <TouchableOpacity style={styles.button} onPress={() => navigation.navigate('Social')}>
+          <Text style={styles.buttonText}>Sign in</Text>
+        </TouchableOpacity>
+      </View>
+    </View>    
+  )
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#182B49',
+    paddingBottom: 40,
+  },
+  title: {
+    fontSize: 48,
+    fontWeight: 'bold',
+    marginBottom: 20,
+    color: 'white',
+  },
+  button: {
+    backgroundColor: '#C69214',
+    paddingTop: 8,
+    paddingRight: 10,
+    paddingBottom: 8,
+    paddingLeft: 10,
+    borderRadius: 8,
+    marginBottom: 30, // Adds spacing between buttons
+    width: 190,
+  },
+  buttonText: {
+    color: 'white',
+    fontSize: 18,
+    fontWeight: 'bold',
+    textAlign: 'center',
+  },
+  textInputLabel: {
+    color: 'white',
+    alignSelf: 'flex-start',
+    fontSize: 16,
+  },
+  textInput: {
+    color: 'white',
+    borderWidth: 1,
+    borderColor: 'white',
+    borderRadius: 5,
+  },
+  login: {
+    width: 190,
+  }
+});
+
+export default LoginScreen;
